+++ changelog.md
@@ -5,6 +5,7 @@
 ### New Features
 - Deck resubmission policy is explicitly stated on submission. This was a frequently-asked question. (#230)
 - `mc!topcut` supports arbitrary size top cuts instead of specifically 8. The size is now required. (#231)
+- `mc!removehost` supports IDs to allow removing hosts who leave the server. (#234)
 
 ## 2021-03-28 ([Chalislime Monthly March 2021](https://challonge.com/csmmar21))